Major change in Safran inputs
The observations are now stored in a tar file for each analysis
execution (the number of obs file in the archive depends on the lead
time).
This allow to run the monthly reanalysis task with only one observation
input file, as for the research reanalysis task.

This modification needed the creation of a new vortex resource
(SafranPackedObs).