Adapt prepsaf task to operational needs
In operational environment the PYTHONPATH does not contain any links to
epygram, vortex or snowtools packages so the 'import' in the makeP
script don't work.
To solve this, the Expresso/ParaExpresso algo components in vortex have
been modified in order to be able to give some path to be temporary
added to the PYTHONPATH environment variable for the execution of the
script. The corresponding footprint is named 'extendedpypath' and has to
be added to the operational tasks.
The path to the modules is taken from links the rootapp path defined in
the job.