Make missing guess files for monthly reanalysis
Task made especially to regenerate guess over Pyrenees as they were not
made over the good grib points. It should not be used for every monthly
reanalysis as the P files must be extracted daily.

There are still many problems with this task :

- As we look for old grib ARPEGE/PEARP file it is necessary to do some
prestaging on hendrix.

- There are missing grib files on hendrix (especially for ARPEGE since
it does work with Vortex yet).

- It is not really efficient, and it had been necessary to launch 1
job/month.