Add flexibility in reanalysis configurations
This commit aims at make it easier to switch from the classic reanalysis
to the configuration using ARPEGE precipitations.

The switch betwwen the 2 configurations is configured with the job-name
that initialise different configuration parameters.

At the moment 2 different laucher are defined, the next step is to make
a launcher script.