Adapt prepsafran tasks to the new HPC environment
The main evolution is the repartition of input files in personnal
subdirectories. It allows to optimise the parallelisation of the
computing part on the 120 available cores without any assumption
on the number of files to deal with.

The PEARP rescue files for the analysis task have also been deleted
since they are already produced in the forecast task.

On form the tasks have been simplified by using the footprints build-in
loops on both members and terms.

The only remaining downside is the fact that ARPEGE and PEARP files
produced by the current operationnal chain on prolix are
stored differently on the HPC than on hendrix (with a filtername
'concatenate' in one case and not in the other). As a consequence the
vortex multi store doesn't work as exepected and 2 different toolbox
with 2 namespace are necessary.
This issue has been submited at IGA team...