Add rescue mode to update the SAFRAN's guess pack for monthly reanalysis
If guesses produced from the assimilation cycle have not been produced,
the corresponding guesses produced with the forecast cycle can be added
to the tar archive that contains all guesses from august 1st.

A comment to prescribe further modification was also added :
since the monthly reanalysis covers the period 01/08 to J-4, the
guess archive only contains guesses up to J-4. It would be more suitable
to add guesses up to J-1, which would allow to have rescue archives.