Restore user-independent environment
This commit reverse a previous one (commit number c6f9cc954c3ebd86ab5a673d653cbba34b14f9f0).

The environment managment in a Vortex task is set up by the jobassistant and its plugins.
But in the case of a 'Expresso' or 'ParaExpresso' execution of a python script the
Vortex-defined environment is lost and only user defined variables (PYTHONPATH) remains.

One Vortex rule is to be user independent : using the PYTHONPATH variable could lead to
use bad versions of the python modules (the default ones instead of the ones bounded to the
actual experiment/vapp/vconf).

The best solution is thus to use the 'extendpypath' footprint of the Expresso/ParaExpresso
algo components to set the script environment with a similar approach than the one use by
the Vortex toolbox.
The major issue is that the root directory of the experiment (old 'rootapp' variable) is no
longer stored by the default Vortex job templates. The idea here is to retrieve this path
from the configuration file path stored in the 'iniconf' configuration variable and that is
always $rootapp/conf/conf_name.ini.