Add documentation to snowtools project in doc/
Add a doc/ folder and a first skeleton for generating snowtools
documentation with sphinx.

Some files of the snowtools projects have been modified to include more
documentation and serve as an example.

As sphinx is run with pyhton3 and all modules are loaded independantly,
some files have been modified to allow python3 loading (print with
parenthesis and absolute imports).