read lat and lon and station from GRID.nc for interpolation on a 1D grid and write it to output file. Issue #2030.
The station variable is needed in the output file when it is used for crocus simulations at a specific location with a terrain shading mask. In the case of a 1D file, it is more intuitive to have the lon and lat of the "pseudo station" in the output file than the lat and lon of the massif centroid, which was the old behaviour.