DBUG archive of guess files in refill task
The identification of guess files to archive using a "guess.out" file
works only when the guess file has been generated by the task but not
when the guess files was already in the Vortex cache.
The new way to loop over guess files to archive is more consistent with
the way the guess files are checked in the first place.