Add an s2m option to add extra inputs to a task
Context :
---------
Giulia's experiments require the definition of different
vegetation types coming from several user-defined ascii files.

Description of the evolution :
------------------------------

This commit introduces a --uenv option to the s2m command for the user
to specify a directory containing extra files (not among the "standard"
SURFEX inputs) for its experiment.
This option works for both local and Vortex/HPC simulations.

* For a local execution, the files are simply copied into the work
  directory

* For a Vortex execution :
A Vortex-like Uenv with the files provided along with the --uenv option
is created of updated.

A class manages the Uenv generation (as well as its extension
for future usage), adding each file under the --uenv directory into a
proper uenv environment (copy files under user's hack data directory and
add a env "dictionnary" file under the user's hack env directory).
A versionning is done for all file that differ from the previous
execution.

This class returns the list of valid entries of the uenv. This list, as
well as the uenv cycle name are writen in the task's configuration file
to be used by the surfex_task.
All entries returned by the UserEnv class are retrieved in the task's
working directory.

A "README" file is created/appended in the directory where the user
stores its files to provide information on the associated UEnvs.

A control is applied on the size and number of files to avoid problems.

Constraints :
-------------
- The uenv keys are used in the surfex_task task to define
the local name of the ressources, so the uenv must not be manually
modified by the user (or very carrefully).
- The user must ensure that the directory provided in the --uenv
option contains all and only the relevant files to avoid useless
transfers.