EVOLUTION Standardisation of Vortex tasks
Context:
--------
The construction of the EDELWEISS system in the coming month/years will
require the development of new Vortex tasks. This evolution introduces
a re-organisation of Vortex tasks managment to make the development of
new tasks easier. The goal is to put all vortex-specific elements in the
background and use only user-readable syntax in the core of the task.

Description of the evolution:
-----------------------------

- An abstract `VortexTask` is introduced to manage all common elements
  of the various Vortex tasks. The different steps are called in methods
  with more explicit names (ex: get_remote_inputs instead of
  'early-fetch') that are to be overwritten for the specific needs of
  each task

- All toolbox.input [/output] calls are done in the `vortexIO` module to
  avoid code duplicates. An input call thus looks like "get_forcing(...)".

TODO :
------
- manage alternates in the vortexIO module
- standardisation of "static" resources in the vortexIO module
- the algo component should also be called in an externalised
  module.