DBUG Use Vortex's job launcher member detection tool
The '_mkjob_opts_detect_2' function of vortex/src/vortex/layout/jobs.py
converts "_mbX" job names extensions into a proper ensemble member.

The idea is to keep a single 'job' section for all the jobs associated
to an ensemble of simulations in the configuration file.

Example :
---------
Let's consider that we want to launch an execution of
SURFEX with an MPI parallelisation for every FORCING file of an N
members ensemble.
We want all these jobs to share a common 'surfex_mpi' section in the
configuration file.

Naming the jobs 'surfex_mpi_mb1', 'surfex_mpi_mb2', ...,  'surfex_mpi_mbN'
will lead the Vortex's job launcher to do this :
- Get each job 'member' from the mbX extension : member=int(X)
- put the associated value direcly in the job (variable RD_MEMBER in the
  template file, then interpreted as 'member' in the configuration
  dictionary)
- rename all jobs 'surfex_mpi' so that they all use the same section of
  the configuration file