Rename script and return Nan 'scd' instead of 0 for seasons/points with no snow
Snow covered days depend on a simulated snow depth threshold (20cm by
default). When the threshold is never reached for a specific point /
season, the previous version of the code returned a melt out date of '0
days since september 1st' but in this case the expected value is NaN.

This commit simply returns a Nan snow melt out date value for season /
points where the computed snow cover duration is 0, which is equivalent
to the condition "no days with snow depth above threshold".