Add a script to clean resources from the local Vortex cache
This is usefull for situations when a resources has already been
downloaded once and is still on the local cache (for exampe on sxcen)
but it has been updated on hendrix (from a HPC simulaton for example).
In this case we want to ensure that downstream scripts use the updated
version of the resource and not the old one (which Vortex will take by
default).

This scripts allows to clean the relevant resource(s) from the local
cache. It can also be called directly from the downstream script.