add crps.f90 and switch to meson as a build backend.
Why switch to meson ?
setuptools does not support extensions written in Fortran and there is no
plan to implement Fortran support in setuptools.
Until numpy<1.26 and python<3.12 Fortran compilation is supported via a numpy layer on top of setuptools.distutils.
For recent versions of numpy and ultimately from python 3.12 on, numpy.f2py relies on meson as a build backend. This means that meson is needed anyway when using numpy.f2py to compile extensions written in Fortran.
The only alternative build backend supporting Fortran would be CMake.
Advantages of meson:
* supports editable installs even for compiled extensions.
* examples are available due to a bunch of projects within the scientific python universe recently switched to meson (scipy, numpy, pandas, scikit-image, scikit-fmm...)
* more readable and easier to understand build instruction files than CMake
Disadvantage:
* not installed system-wide on our computers. (Not a problem when working with virtual environments and access to pypi.)