Uniformisation of members management in the configuration file
The crocO task requires the member variable in the configuration
file to be a "rangex(start:${start} end:${end})" string, while the
surfex_mpi task requires to loop over forcing ensemble members (needing
to retrive the list from the "ensemble" options/conf variable).

This commit generalise the use of the "rangex" syntax for member
variables and uses regular expression to retrive the list of members for
surfex_mpi applications.