Tests: Cleaning of temporary folder creation
Creation of tool classes to change to the temporary folder.

Use of this tool in test bdap and hydro to ensure the coming back to the
(existiong) previous folder is not forgotten

Logging of output of subprocess.run (that is replaced with run_and_log
class function).

Do not wreite into the repository (fail_test folder) but let temfolder
choose the correct temp folder to use (generally /tmp or similar folder)
to prevent mixing of code and output of tests.