DBUG. Disable tool if rioxarray is unvailable + ensure proper spatial dimension names
The snowtools grid re-projection method is based on rioxarray, which
is not currently available on default installations.
The availablity of rioxarray is now checked with bronx's ExternalCodeImportChecker
tool.

A mapping between the spatial dimension names used by default in snowtools (xx, yy) and
by rioxarray (x, y) is also done to ensure that:
- rioxarray does not crash
- the "proj" method returns an xarray object with the valid (xx, yy)
  dimensions