Dev: Adapt xaray devs to HPC
- The xarray backend module is not available on HPC, so it must be
disabled in this case and the "preprocess" method must be called
directly. To do this, a clear split between the class inheriting from xarray
backend and the preprocess methods is suitable so a new
xarray_snowtools_preprocess module has been added containing all
preprocess methods that can be called directly on HPC or through the
snowtools backend when available.

- typing.Literal and simplified Union notation is not available on HPC.