DBUG: xarray_snowtools_backend - cleaning and nsure that a closing method is defined
The "open_dataset" method alone should be defined in SnowtoolsBackendEntrypoint
because the "open_mfdataset" and "open_datarray" methods of xarray
ultimately call the open_dataset method with the proper engine.

The preprocessing step generates a new dataset object without the "_close"
attribute of the original object, which disables the closing of opened
files. To ensure a better behavior, the closing method of the original
object is saved before the preprocessing step to be re-used as the
closing method in the pre-processed dataset.