DBUG preprocess backtracking
- Save the mapping dictionnary to original dimension/variable names
as type "str" instead of "dict".
This is necessary to avoid issues when writing data in a NetCDF file
beacuse the serialization to netCDF files is not possible with dictionnary-like attributes.

Thus, the xarray native "to_netcdf" methods rises a TypeError with the
following message:
TypeError: Invalid value for attr 'original_variable_name': {'massif_num': 'massif_number'}.
For serialization to netCDF files, its value must be of one of the following types:
str, Number, ndarray, number, list, tuple

- Combine the "update_varname" and "update_dimname" methods into a single
"update_names" method to be consistent with the xarray native "rename"
method.